Split out utility functions and constants to new file
